fixed service ports and container ports
## Renamed `http` ports to `web`

`http` is not recommended as a port name with custom values as its a valid recognized port for k8s as this is a IANA standard port.

## Exposed 80 (HTTP) ports on k8s services

There's no need to configure k8s service ports to some arbitrary values. It's better using 80 (http) port on services, and keeping arbitrary values only for container ports.

This way services can be accessed from within the cluster using the service name instead of hardcoding the full k8s resource name around